django-event-management/requirements.txt
asgiref==3.2.10
bson==0.5.8
dataclasses==0.6
Django==2.2.13
django-betterforms==1.2
django-ckeditor==5.9.0
django-crispy-forms==1.9.1
django-js-asset==1.2.2
django-mapbox-location-field==1.5.0
djongo==1.3.2
dnspython==1.16.0
# Pillow==7.1.2
Pillow==9.2.0
pymongo==3.10.1
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.2.4
psycopg2==2.8.6